[REFACTOR] Move formatting utilities to own file
