+++ pandora-common/src/utility/index.ts
@@ -1,1 +1,2 @@
+export * from './formatting';
 export * from './misc';
